`jellyfish-wallet-encrypted` to use `scrypt-js`, a pure js scrypt implementation (#531)
* Revert "refactor `scryptsy` into jellyfish-wallet-encrypted/test (#516)"

This reverts commit 053839f211d7d674ebd8c291c7c140962176dd27.

* refactor and add scrypt-js
