+++ packages/jellyfish-wallet-encrypted/__tests__/scrypt.test.ts
@@ -1,84 +1,36 @@
 import { Scrypt } from '../src'
-import { SimpleScryptsy } from './SimpleScryptsy'
 
-const scrypt = new Scrypt(new SimpleScryptsy())
+it('should be able to generate from random (same) passphrase into multiple desired length secret', async () => {
+  const scryptsy = new Scrypt()
+  const secret1 = await scryptsy.derive('random password', Buffer.from('abcd', 'hex'), 64)
+  expect(secret1.length).toStrictEqual(64)
 
-it('should be able to encrypt / decrypt', async () => {
-  const privKey = 'e9873d79c6d87dc0fb6a5778633389f4e93213303da61f20bd67fc233aa33262'
-  const passphrase = 'password'
-  const buffer = Buffer.from(privKey, 'hex')
-
-  const data = await scrypt.encrypt(buffer, passphrase)
-  expect(data).not.toStrictEqual(null)
-  expect(data.encode()).not.toStrictEqual(null)
-
-  const encoded = data.encode()
-  const decrypted = await scrypt.decrypt(encoded, passphrase)
-  expect(decrypted).toStrictEqual(buffer)
-  expect(decrypted.toString('hex')).toStrictEqual(privKey)
+  const secret2 = await scryptsy.derive('random password', Buffer.from('abcd', 'hex'), 40)
+  expect(secret2.length).toStrictEqual(40)
 })
 
-it('should be able to encrypt / decrypt - simple passphrase, a 6 digit pin', async () => {
-  const privKey = 'e9873d79c6d87dc0fb6a5778633389f4e93213303da61f20bd67fc233aa33262'
-  const passphrase = '135790'
-  const buffer = Buffer.from(privKey, 'hex')
+it('configurable params (easy-hard)', async () => {
+  const easy = new Scrypt(
+    16384,
+    8,
+    1
+  )
 
-  const data = await scrypt.encrypt(buffer, passphrase)
-  expect(data).not.toStrictEqual(null)
-  expect(data.encode()).not.toStrictEqual(null)
+  const hard = new Scrypt(
+    16384,
+    8,
+    8
+  )
 
-  const encoded = await data.encode()
-  const decrypted = await scrypt.decrypt(encoded, passphrase)
-  expect(decrypted).toStrictEqual(buffer)
-  expect(decrypted.toString('hex')).toStrictEqual(privKey)
-})
+  const easyStart = Date.now()
+  await easy.derive('random password', Buffer.from('abcd', 'hex'), 32)
+  const easyTime = Date.now() - easyStart
 
-it('Should work with variable data length - long', async () => {
-  const privKey = 'e9873d79c6d87dc0fb6a5778633389f4e93213303da61f20bd67fc233aa33262e9873d79c6d87dc0fb6a5778633389f4e93213303da61f20bd67fc233aa33262e9873d79c6d87dc0fb6a5778633389f4e93213303da61f20bd67fc233aa33262e9873d79c6d87dc0fb6a5778633389f4e93213303da61f20bd67fc233aa33262'
-  const passphrase = 'passcode'
-  const buffer = Buffer.from(privKey, 'hex')
+  const hardStart = Date.now()
+  await hard.derive('random password', Buffer.from('abcd', 'hex'), 32)
+  const hardTime = Date.now() - hardStart
 
-  const data = await scrypt.encrypt(buffer, passphrase)
-  expect(data).not.toStrictEqual(null)
-  expect(data.encode()).not.toStrictEqual(null)
-
-  const encoded = data.encode()
-  const decrypted = await scrypt.decrypt(encoded, passphrase)
-  expect(decrypted).toStrictEqual(buffer)
-  expect(decrypted.toString('hex')).toStrictEqual(privKey)
-})
-
-it('should work with variable data length - short', async () => {
-  const privKey = '1234'
-  const passphrase = 'passcode'
-  const buffer = Buffer.from(privKey, 'hex')
-
-  const data = await scrypt.encrypt(buffer, passphrase)
-  expect(data).not.toStrictEqual(null)
-  expect(data.encode()).not.toStrictEqual(null)
-
-  const encoded = data.encode()
-  const decrypted = await scrypt.decrypt(encoded, passphrase)
-  expect(decrypted).toStrictEqual(buffer)
-  expect(decrypted.toString('hex')).toStrictEqual(privKey)
-})
-
-it('should accept "even" number data length', async () => {
-  const buffer = Buffer.from('eeffaabb', 'hex')
-
-  const data = await scrypt.encrypt(buffer, 'password')
-  expect(data).not.toStrictEqual(null)
-})
-
-it('should reject "odd" number data length', async () => {
-  const buffer = Buffer.from('eeffaa', 'hex')
-
-  await expect(async () => {
-    await scrypt.encrypt(buffer, 'password')
-  }).rejects.toThrow('Data length must be even number')
-})
-
-it('should pad "odd" number data length', async () => {
-  const padded = Buffer.from('00' + 'eeffaa', 'hex')
-  await scrypt.encrypt(padded, 'password')
+  // significantly slower
+  // technically it is 8x harder, but they can be processed in parallel
+  expect(hardTime).toBeGreaterThan(easyTime * 2)
 })
