+++ packages/jellyfish-wallet-encrypted/src/index.ts
@@ -1,3 +1,3 @@
+export * from './encryption'
 export * from './hd_node'
-export * from './provider/scrypt_provider'
 export * from './scrypt'
