--- packages/jellyfish-wallet-encrypted/src/provider/scrypt_provider.ts
@@ -1,3 +1,0 @@
-export interface ScryptProvider {
-  passphraseToKey: (nfcUtf8: string, salt: Buffer, desiredKeyLen: number) => Promise<Buffer>
-}
